Write each module functionality and import them to index.js
